refactor:communityApi폴더에 커뮤니티 관련 api함수 정리
